app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
pytesseract
opencv-python-headless
pillow
python-decouple
langchain
numpy